+++ reportCliError.test.mjs
@@ -4,7 +4,7 @@ import { strictEqual, throws } from "node:assert";
 import { spawnSync } from "node:child_process";
 import { fileURLToPath } from "node:url";
 import replaceStackTraces from "replace-stack-traces";
-import snapshot from "snapshot-assertion";
+import assertSnapshot from "snapshot-assertion";
 
 import reportCliError from "./reportCliError.mjs";
 
@@ -51,7 +51,7 @@ export default (tests) => {
 
       strictEqual(stdout.toString(), "");
 
-      await snapshot(
+      await assertSnapshot(
         replaceStackTraces(stderr.toString()),
         new URL(
           "./test/snapshots/reportCliError/Error-instance-with-stack-stderr.ans",
@@ -88,7 +88,7 @@ export default (tests) => {
 
       strictEqual(stdout.toString(), "");
 
-      await snapshot(
+      await assertSnapshot(
         replaceStackTraces(stderr.toString()),
         new URL(
           "./test/snapshots/reportCliError/Error-instance-without-stack-stderr.ans",
@@ -123,7 +123,7 @@ export default (tests) => {
 
     strictEqual(stdout.toString(), "");
 
-    await snapshot(
+    await assertSnapshot(
       replaceStackTraces(stderr.toString()),
       new URL(
         "./test/snapshots/reportCliError/primitive-value-stderr.ans",